Rename from command to service
